taint/test/mochitest/test_non_string_sinks.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test HTML Message Port Taint Sink With Non String</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      let number_as_string = "10";
      // The number of flows produced is explained later on
      let number_of_tainted_flows = 14;

      let i = 0;
      
      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {
          is(report.detail.str, Number(number_as_string), "Check sink string content");
          i += 1;
          if (i >= number_of_tainted_flows) {
            SimpleTest.finish();
          }
      }, false);

      // Worker script as a Blob
      const workerScript = `
          self.onmessage = (event) => {
            if (event.data === 'initialize') {
                // Retrieve the port from the message
                const port = event.ports[0];

                // Listen for messages on the port
                port.onmessage = (event) => {
                    console.log(JSON.stringify(event));
                    port.postMessage("Received the event");
                };

                // Send an initial message back to the main page
                port.postMessage('Worker initialized and ready.');
            }
        };
      `;
      if (window.Worker) {
        const blob = new Blob([workerScript], { type: 'application/javascript' });
          const workerUrl = URL.createObjectURL(blob);

          const worker = new Worker(workerUrl);

          const channel = new MessageChannel();

          // Send one of the ports to the worker
          worker.postMessage('initialize', [channel.port1]);

          // Listen for messages from the worker
          channel.port2.onmessage = (event) => {
              console.log('Message received from worker:', event.data);
          };

          // Should produce 1 taint report
          let tainted_number = Number(String.tainted(number_as_string));
          channel.port2.postMessage(tainted_number);

          // Should produce 2 taint report
          let tainted_object = {x:tainted_number, y:tainted_number};
          channel.port2.postMessage(tainted_object);

          // Should produce 4 taint report
          let tainted_complex_object = {x:tainted_number, y:tainted_object, z:tainted_number};
          channel.port2.postMessage(tainted_complex_object);

          // Should produce 7 taint report
          let tainted_array = [tainted_number, tainted_complex_object, tainted_object];
          channel.port2.postMessage(tainted_array);
        } else {
            console.error('No support for workers');
        }
    </script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <p id="test"></p>
    <button id="btn"></button>
  </body>
</html>
